Replace eslint and prettier with biome
