--- .eslintignore
@@ -1,11 +1,0 @@
-node_modules
-
-# contains no src files
-packages/hyperdrive-artifacts
-
-# generated files
-packages/hyperdrive-wasm
-packages/fixed-point-wasm
-
-# sandbox
-apps/sdk-sandbox
