--- apps/hyperdrive-sdk-docs/.eslintrc
@@ -1,20 +1,0 @@
-{
-  "extends": [
-    "plugin:react/recommended",
-    "plugin:react/jsx-runtime",
-    "plugin:@docusaurus/recommended",
-    "@hyperdrive/eslint-config"
-  ],
-  "parserOptions": {
-    "sourceType": "module"
-  },
-  "rules": {
-    "@typescript-eslint/no-var-requires": "off",
-    "import/no-unresolved": [
-      2,
-      {
-        "ignore": ["^@theme", "^@docusaurus", "^@site"]
-      }
-    ]
-  }
-}
