--- apps/hyperdrive-sdk-docs/.prettierignore
@@ -1,6 +1,0 @@
-build
-node_modules
-.docusaurus
-
-# generated files
-docs/sdk/api-reference
