--- apps/hyperdrive-trading/.eslintignore
@@ -1,5 +1,0 @@
-node_modules
-src/types
-dist
-.prettierrc.js
-settings.json
