--- apps/hyperdrive-trading/.eslintrc
@@ -1,23 +1,0 @@
-{
-  "root": true,
-  "extends": [
-    "plugin:react/recommended",
-    "plugin:react-hooks/recommended",
-    // Removes the requirement to import `React` when you use jsx
-    "plugin:react/jsx-runtime",
-    "plugin:tailwindcss/recommended",
-    "plugin:@tanstack/eslint-plugin-query/recommended",
-    // @hyperdrive/eslint-config needs to be the last config extended because it includes prettier
-    "@hyperdrive/eslint-config",
-  ],
-  "plugins": ["react-refresh"],
-  "rules": {
-    // Disable ordering in favor of prettier plugin
-    "tailwindcss/classnames-order": "off",
-    "tailwindcss/no-custom-classname": "off",
-    // Exported functions should be written as function declarations for easier
-    // scanning. Arrow functions can still be used as anonymous functions.
-    "func-style": ["error", "declaration"],
-    "react-refresh/only-export-components": "warn",
-  },
-}
