+++ apps/hyperdrive-trading/package.json
@@ -15,9 +15,6 @@
   "scripts": {
     "build": "bash ./scripts/build-app.sh",
     "dev": "vite",
-    "format": "prettier --write '**/*.{gql,graphql,js,jsx,ts,tsx,json,md,yaml,yml}'",
-    "format:check": "prettier --check '**/*.{gql,graphql,js,jsx,ts,tsx,json,md}'",
-    "lint": "eslint src",
     "serve": "vite preview",
     "test": "vitest run",
     "test:watch": "vitest --reporter=verbose",
@@ -61,28 +58,16 @@
     "wagmi": "^2.5.12"
   },
   "devDependencies": {
-    "@hyperdrive/eslint-config": "*",
-    "@hyperdrive/prettier-config": "*",
     "@hyperdrive/tsconfig": "*",
-    "@tanstack/eslint-plugin-query": "^5.28.11",
     "@tanstack/router-devtools": "^1.9.0",
     "@tanstack/router-vite-plugin": "^1.8.2",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.10",
     "@types/react-helmet": "^6.1.6",
-    "@typescript-eslint/eslint-plugin": "^7.5.0",
-    "@typescript-eslint/parser": "^7.5.0",
     "@vitejs/plugin-react": "^3.1.0",
     "autoprefixer": "^10.4.13",
-    "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.9",
-    "eslint-plugin-tailwindcss": "^3.15.1",
     "msw": "^2.3.1",
     "postcss": "^8.4.21",
-    "prettier": "3.3.3",
-    "prettier-plugin-organize-imports": "4.0.0",
-    "prettier-plugin-tailwindcss": "^0.6.5",
     "tailwindcss": "^3.4.7",
     "tailwindcss-debug-screens": "2.2.1",
     "typescript": "^5.5.4",
