--- packages/eslint-config/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@hyperdrive/eslint-config",
-  "version": "0.0.0",
-  "private": true,
-  "main": ".eslintrc",
-  "type": "module",
-  "license": "AGPL-3.0",
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.5.0",
-    "@typescript-eslint/parser": "^7.5.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-config-turbo": "^2.0.6",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-unicorn": "^52.0.0",
-    "typescript": "^5.4.4"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
