--- packages/gopher/.eslintrc
@@ -1,4 +1,0 @@
-{
-  "root": true,
-  "extends": ["@hyperdrive/eslint-config"]
-}
