--- packages/hyperdrive-appconfig/.eslintrc
@@ -1,10 +1,0 @@
-{
-  "root": true,
-  "extends": [
-    "@hyperdrive/eslint-config"
-  ],
-  "rules": {
-    "turbo/no-undeclared-env-vars": "off",
-    "no-console": "off"
-  }
-}
