+++ packages/hyperdrive-appconfig/package.json
@@ -6,9 +6,7 @@
   "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "files": ["dist"],
   "scripts": {
     "gen:local": "npx tsx ./src/scripts/local.ts",
     "gen:cloudchain": "npx tsx ./src/scripts/cloudChain.ts",
@@ -20,15 +18,12 @@
   "devDependencies": {
     "@delvtech/hyperdrive-artifacts": "^1.0.16",
     "@delvtech/hyperdrive-viem": "^2.4.3",
-    "@hyperdrive/eslint-config": "*",
-    "@hyperdrive/prettier-config": "*",
     "@hyperdrive/tsconfig": "*",
     "lodash.uniqby": "^4.7.0",
     "@types/lodash.uniqby": "^4.7.9",
     "abitype": "^0.9.8",
     "dotenv": "^16.0.3",
     "tsx": "^4.7.0",
-    "prettier-plugin-organize-imports": "3.2.4",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.1.0",
     "tsup": "^7.2.0",
