--- packages/hyperdrive-artifacts/.eslintignore
@@ -1,3 +1,0 @@
-node_modules
-hyperdrive-temp.*
-dist
