--- packages/hyperdrive-artifacts/.prettierignore
@@ -1,3 +1,0 @@
-node_modules
-hyperdrive
-!dist
