--- packages/hyperdrive-js-core/.prettierignore
@@ -1,3 +1,0 @@
-node_modules
-src/types
-dist
