+++ packages/hyperdrive-js-core/package.json
@@ -22,16 +22,12 @@
   },
   "devDependencies": {
     "@delvtech/evm-client": "^0.5.1",
-    "@hyperdrive/eslint-config": "*",
-    "@hyperdrive/prettier-config": "*",
     "@hyperdrive/tsconfig": "*",
     "@types/lodash.groupby": "^4.6.9",
     "@types/lodash.mapvalues": "^4.6.9",
     "@types/sinon": "^17.0.3",
     "abitype": "^1.0.5",
     "dotenv": "^16.4.5",
-    "prettier": "3.3.3",
-    "prettier-plugin-organize-imports": "4.0.0",
     "sinon": "^18.0.0",
     "tsconfig-paths": "^4.2.0",
     "tsup": "^8.2.3",
@@ -41,9 +37,7 @@
   },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "files": ["dist"],
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
