--- packages/hyperdrive-viem/.prettierrc.mjs
@@ -1,7 +1,0 @@
-import repoConfig from "@hyperdrive/prettier-config";
-
-/** @type {import("prettier").Config} */
-export default {
-  ...repoConfig,
-  plugins: ["prettier-plugin-organize-imports"],
-};
