+++ packages/hyperdrive-viem/package.json
@@ -16,19 +16,14 @@
     "@delvtech/evm-client-viem": "^0.6.1"
   },
   "devDependencies": {
-    "@hyperdrive/eslint-config": "*",
-    "@hyperdrive/prettier-config": "*",
     "@hyperdrive/tsconfig": "*",
-    "prettier-plugin-organize-imports": "3.2.4",
     "tsup": "^7.2.0",
     "typescript": "^5.0.2",
     "viem": "^2.7.8"
   },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "files": ["dist"],
   "exports": {
     ".": {
       "import": "./dist/index.js",
