--- packages/prettier-config/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": ["@hyperdrive/eslint-config"]
-}
