--- packages/prettier-config/.prettierrc.js
@@ -1,11 +1,0 @@
-/** @type {import("prettier").Config} */
-const config = {
-  tabWidth: 2,
-  useTabs: false,
-  trailingComma: "all",
-  singleQuote: false,
-  semi: true,
-  printWidth: 80,
-};
-
-export default config;
