--- packages/prettier-config/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@hyperdrive/prettier-config",
-  "version": "0.0.0",
-  "private": true,
-  "main": ".prettierrc.js",
-  "license": "AGPL-3.0",
-  "type": "module",
-  "dependencies": {
-    "prettier": "3.2.5"
-  },
-  "scripts": {
-    "lint:w": "eslint --fix '**/*.{gql,graphql,js,jsx,ts,tsx,json,md}'",
-    "lint": "eslint '**/*.{gql,graphql,js,jsx,ts,tsx,json,md}'",
-    "format": "prettier --write '**/*.{gql,graphql,js,jsx,ts,tsx,json,md}'",
-    "format:check": "prettier --check '**/*.{gql,graphql,js,jsx,ts,tsx,json,md}'"
-  }
-}
