--- packages/tsconfig/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": ["@council/eslint-config"]
-}
