+++ packages/tsconfig/package.json
@@ -2,12 +2,5 @@
   "name": "@hyperdrive/tsconfig",
   "version": "0.0.0",
   "private": true,
-  "files": [
-    "base.json",
-    "nextjs.json",
-    "react-library.json"
-  ],
-  "devDependencies": {
-    "@hyperdrive/eslint-config": "*"
-  }
+  "files": ["base.json", "nextjs.json", "react-library.json"]
 }
